# Faster TreeOfLosers