Simplifies TreeOfLosers to work with raw pointers to each merge source. Encodes
a tree as an array of indices into the array of sources. The levels of a binary
tree are stored consecutively. The index of the left child of node i is i * 2 +
1 and the right child is at i * 2 + 2.

Implements the tree of losers algorithm iteratively. The initialization fills
the tree with values and returns the first value. The value at each node is the
greater value of its two children or kEmpty if the children are at end. The
next function pops a value off the previous winner and reinserts the winner at
in the parent of the leaf node representing the stream. The value goes up the
tree and if it wins and stays at the parent where it loses and the winner
(previous value of the node) continues upward.

Adds a class  MergeStream to document the interface expected from mergeable
sources.

Adds an alternate implementation of merge as a sorted array of sources. When
popping a value off a source, reinserts the array in the place corresponding to
the new first element and shifts the sources below the insertion point.

Adds a benchmark comparing the tree of losers to the sorted array of streams.
They are closely matched at small stream counts and the tree is better at large
counts.

```
narrowTree                                                    2.02s  493.94m
narrowArray                                      110.58%      1.83s  546.18m
mediumTree                                                    3.40s  294.27m
mediumArray                                      103.51%      3.28s  304.58m
wideTree                                                      6.73s  148.60m
wideArray                                         47.53%     14.16s   70.62m
```

Narrow is 7 streams, medium is 37 and wide is  1029 streams.