@SergeyLavr 2022-03-25
My question is in the context of LocalMerge, where @mbasmanova proposed to try TreeOfLosers to improve performance. In earlier version LocalMerge was using priority_queue on top of vector. As I understand, is implemented as Heap, which for small number of elements can have good performance due to locality. However, standard priority_queue have resulted in excessive copy operations due to details of stl implementation and overall performance was bad.
Is there a comparison of performance between Heap and TreeOfLosers? I wonder it is worth to have own-implementation of Heap and use it vs. TreeOfLosers depending on the context.