@oerling 2022-03-26
The tree of losers could be flattened into an array, which would make less  dirty cache lines. Like a struct or arrays. 

My guess is that this could be made a fixed fanout tree, like 16 children per node. Sort the children within the node, then pop the least one, then reinsert the popped one according to the next element. For cases of more than 16 streams, we get a tree with the tree of losers logic of pulling from the child with the lowest value.  Shifting items around to do insertion sort in a small array is more local then the tree operations of tree of losers or heap butt has lots of instructions to either loop or switch on the width and use wider loads and stores. 

 

If this goes under 10% of the profile, it will not be worth much more work.

 

If this is over 10% of some reasonable operation then it could make sense to flatten the data structure and try a arity higher than binary. Let’s see how this works in the merge application and then maybe flatten and increase arity.

 

 

From: Sergii Lavrinenko ***@***.***> 
Sent: Friday, March 25, 2022 2:55 PM
To: facebookincubator/velox ***@***.***>
Cc: oerling ***@***.***>; Author ***@***.***>
Subject: Re: [facebookincubator/velox] Faster TreeOfLosers (PR #1289)

 

My question is in the context of LocalMerge, where @mbasmanova <https://github.com/mbasmanova>  proposed to try TreeOfLosers to improve performance. In earlier version LocalMerge was using priority_queue on top of vector. As I understand, is implemented as Heap, which for small number of elements can have good performance due to locality. However, standard priority_queue have resulted in excessive copy operations due to details of stl implementation and overall performance was bad.
Is there a comparison of performance between Heap and TreeOfLosers? I wonder it is worth to have own-implementation of Heap and use it vs. TreeOfLosers depending on the context.

—
Reply to this email directly, view it on GitHub <https://github.com/facebookincubator/velox/pull/1289#issuecomment-1079464703> , or unsubscribe <https://github.com/notifications/unsubscribe-auth/AKPPPT7ITIMRBIPRT3OKRSLVBYYZ7ANCNFSM5RTA6MRQ> .
You are receiving this because you authored the thread.Message ID: ***@***.***>