@mbasmanova 2022-03-28
@oerling Orri, I see "This branch has conflicts that must be resolved" on this PR. Would you rebase and update? It might be helpful to add a benchmark for the ToL.