@oerling 2022-03-29
I  updated the algorithm one more time. Also added one that just does a binary search and shifts the streams. The shifting is a little faster < 10% for few streams. The tree is better for more streams. The tree comes down pretty much to the heap algorithm. Insert at the leaf, then  go up with max log compares. The tree is represented as an array of stream indices.